+++ classes/users/wpdk-user.php
@@ -679,7 +679,8 @@ class WPDKUsers {
 
     /* Do a several action/filter to monitoring user action. */
 
-    $this->logout();
+    //$this->logout();
+    add_action( 'init', array( $this, 'logout' ) );
 
     /* Main hook for common check in front end. */
     //add_action( 'wp_head', array( $this, 'wp_head_signin' ) );
@@ -715,17 +716,13 @@ class WPDKUsers {
     add_filter( 'user_contactmethods', array( $this, 'user_contactmethods' ) );
   }
 
-  // -----------------------------------------------------------------------------------------------------------------
-  // Check if a user is disabled
-  // -----------------------------------------------------------------------------------------------------------------
-
   /**
    * Force an user logout when disabled or if in GET you pass wpdk_logout
    *
    * @brief Logout an user
    */
-  private function logout() {
-
+  public function logout()
+  {
     /* If a user is logged in. */
     if ( is_user_logged_in() ) {
       $user_id = get_current_user_id();
@@ -738,7 +735,7 @@ class WPDKUsers {
       }
 
       /* Manual logout. */
-      if( isset( $_REQUEST['wpdk_logout'] ) ) {
+      if ( isset( $_REQUEST['wpdk_logout'] ) ) {
         $logout = true;
       }
 
@@ -850,32 +847,47 @@ class WPDKUsers {
 
   }
 
-  // -----------------------------------------------------------------------------------------------------------------
+  // -------------------------------------------------------------------------------------------------------------------
   // Signin utility
-  // -----------------------------------------------------------------------------------------------------------------
+  // -------------------------------------------------------------------------------------------------------------------
 
   /**
    * Do a WordPress Sign in and call filters and action
    *
    * @brief Signin
    *
-   * @param string $email    A valid email address
-   * @param string $password Password
-   * @param bool   $remember Optional. TRUE for set a cookie for next login
+   * @param string|int $user     Any user id, user email or user login
+   * @param string     $password Password
+   * @param bool       $remember Optional. TRUE for set a cookie for next login
    *
    * @return bool TRUE if success, FALSE for access denied
    */
-  public function signIn( $email, $password, $remember = false ) {
+  public function signIn( $user, $password, $remember = false ) {
 
-    /* Sanitize email */
-    $email = sanitize_email( $email );
+    $user_id = false;
 
-    /* User exists with email */
-    $id_user = email_exists( $email );
+    /* Check user id */
+    if ( is_numeric( $user ) ) {
+      $user_id = $user;
+    }
+    /* Check for email */
+    elseif ( is_email( $user ) ) {
+      /* Sanitize email */
+      $email = sanitize_email( $user );
+      /* User exists with email */
+      $user_id = email_exists( $email );
+    }
+    /* Check for user login */
+    else {
+      $user = get_user_by( 'login', $user );
+      if( $user ) {
+        $user_id = $user->ID;
+      }
+    }
 
-    if ( false !== $id_user ) {
-      $user = new WP_User( $id_user );
-      if ( $user ) {
+    if ( false !== $user_id ) {
+      $user = new WPDKUser( $user_id );
+      if ( $user->exists() && WPDKUserStatus::DISABLED !== $user->status ) {
         $result = wp_authenticate( $user->user_login, $password );
         if ( !is_wp_error( $result ) ) {
           /* Set remember cookie */
@@ -893,6 +905,7 @@ class WPDKUsers {
     }
 
     do_action( 'wpdk_singin_wrong', $email, $password );
+
     return false;
   }